refactor(docs): lift the colour-write policy out of tier-sync
The hue-gap colour picker (pickNextColor + the RGB/HSL helpers and the
MIN_HUE_GAP constant) was buried inside tier-sync-diagram.tsx. It is
diagram policy — the engine deliberately knows nothing about what a
write looks like — but it is shared policy: any animated propagation
diagram wants the same guarantee that consecutive writes are visibly
distinct.

Move it to components/mdx/colour.ts and have tier-sync import from
there. No behaviour change; pure relocation.
